+++ js/const.js
@@ -1,0 +1,2 @@
+export const OFFERS_COUNT = 10;
+export const COARD_AMOUNT = 5;
